Fix bad backslash escape usage.
Python 3.6 began reporting a DeprecationWarning for the use of
backslash escape sequence where the second character isn't a valid
Python escape sequence.  Unlike C, in Python, the result of this is
both characters (not just the second).

Python 3.12 changed that to a SyntaxWarning, the final step before
making it a SyntaxError.  So, we should fix it now.

First reported as https://github.com/eventable/vobject/issues/164